Don't try to delete a root directory
This should allow to pass a root directory to delete all its content, without throwing any exception once it's done

+++ securedelete-net/SecureDelete.cs
@@ -66,6 +66,13 @@ namespace SecureDelete
                     DeleteFile(file);
                 }
             }
+            
+            if (string.Equals(Path.GetPathRoot(directoryInfo.FullName), directoryInfo.FullName, StringComparison.OrdinalIgnoreCase))
+            {
+                // can't delete a root
+                return;
+            }
+            
             ObfuscateDirectory(directoryInfo);
             directoryInfo.Delete();
         }
@@ -111,6 +118,13 @@ namespace SecureDelete
                     DeleteFileWithoutDriveDetection(file);
                 }
             }
+            
+            if (string.Equals(Path.GetPathRoot(directoryInfo.FullName), directoryInfo.FullName, StringComparison.OrdinalIgnoreCase))
+            {
+                // can't delete a root
+                return;
+            }
+            
             ObfuscateDirectory(directoryInfo);
             directoryInfo.Delete();
         }
